refactor: change employee service feedback average calc logic

+++ src/main/java/br/com/fiap/sprint1/JavaSprint1WhitelabelAPI/dto/employe/EmployeeDetailsWithFeedbackAverageDTO.java
@@ -2,13 +2,12 @@ package br.com.fiap.sprint1.JavaSprint1WhitelabelAPI.dto.employe;
 
 import br.com.fiap.sprint1.JavaSprint1WhitelabelAPI.model.Employee;
 
-import java.math.BigDecimal;
 
 public record EmployeeDetailsWithFeedbackAverageDTO(
         EmployeeDetailsDTO employeeDetails,
-        BigDecimal feedbackAvarage
+        Double feedbackAvarage
 ) {
-    public EmployeeDetailsWithFeedbackAverageDTO(Employee employee, BigDecimal feedbackAvarage) {
+    public EmployeeDetailsWithFeedbackAverageDTO(Employee employee, Double feedbackAvarage) {
         this(
                 new EmployeeDetailsDTO(employee),
                 feedbackAvarage
